// src/app/auth/interfaces/beneficiario-nuevo.interface.ts
export interface BeneficiarioNuevoInterface {
  nombre: string;
  apellido_paterno: string;
  apellido_materno: string;
  curp?: string;
  id_municipio: string;
  id_localidad: string;
  direccion: string;
  tipo_sangre: string;
  alergias: string;
  telefono_emergencia: string;
  comentarios: string;
  id_discapacidad: string;
  url_fotografia: string;
}